Changed date datatype from Time to DateTime. pry.
My tests would show dates of 01/01/2000 and the intended time of
whatever date I used.  Even the query that searched for events in a
given range would show the correct event but when I asked what the date
of that event was, it would show 01/01/2000 even though what was way
outside the range.

In Gemfile: uncommented out jbuilder and added pry.