corrected response for POST.
Previously, I was getting false-positives of my posts returning the data
{'status' : 'ok'} because I was forcing that response in a helper method
in my support folder for my specs.  When I moved the logic to include
that data to the controller, I also was reminded that I didn't need to
return a hash of the object I just posted (according to the instructions of the
code challenge).  So I changed the spec to match my post values to the object
created by that post vs what was returned in the json of what was created,
which I deleted from the response format.